02_Currency_Converter_Agent/pyproject.toml
[project]
name = "02-currency-converter-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "gradio>=5.33.0",
    "openai-agents>=0.0.17",
]
